# Refactor ManipulateNames to use Target