… src/main/scala/firrtl/transforms/RemoveKeywordCollisions.scala
… src/test/scala/firrtlTests/features/LetterCaseTransformSpec.scala