This PR is doing the following things:

- `RemoveKeywordCollisions` is refactored into a more general `ManipulateNames` transform
- The logic used inside `ManipulateNames` is rewritten to use `Target` instead of `Named`
- Control of `ManipulateNames` is enabled via allowlisting and blocklisting of targets
- Two new transforms are added: `LowerCaseNames` and `UpperCaseNames`
- Command line options to lower/upper case all names are added via: `--change-name-case <lower|upper>`
- Improve the peformance of `VerilogRename`

### Generalize `RemoveKeywordCollisiosns` as `ManipulateNames`

The transform `RemoveKeywordCollisions` is generalized into a new transform, `ManipulateNames`, that does generic string manipulation of the form:

```scala
def manipulate: (String, Namespace) => Option[String]
```

This enables keyword removal and a whole host of other name manipulation to be succinctly handled with one transform.

### Use `Target` instead of `Named`

#1592 (reverted in #1651) was a straight generalization of `RemoveKeywordCollisions` as `ManipulateNames`. However, the former is an older transform that uses `Named` to track where it is in the AST. Since `Named` is being replaced by `Target`, this PR also rewrites the logic inside `ManipulateNames` to use `Target` exclusively.

This process resulted in bug fixes in name manipulation that crop up. See the modifications to `VerilogEmtiterTests.scala` for an example of a test that fails with the old logic.

### Control `ManipulateNames` via allowlisting and blocklisting, skip `ir.ExtModule`

Users doing name manipulation likely do not want to blindly rename everything. First, all `ir.ExtModule`s are skipped. These are external interfaces and can't be messed with. Second, annotations for controlling what is renamed in a specific sub-class of `ManipulateNames` are provided:

- `ManipulateNamesBlocklistAnnotation` to skip targets in the circuit
- `ManipulateNamesAllowlistAnnotation` to only rename specific targets

Names in the circuit will be renamed if they are not on the blacklist and they are on the allowlist. There are a few deviations from this related to the circuit and the top module:

- blocklisting the circuit means skipping the circuit entirely
- if an allowlist is used the circuit must be allowlisted
- if the top module is renamed the circuit will be renamed

### Letter Case Transforms

Two new sub-classes of `ManipulateNames` are added:

- `firrtl.features.LowerCaseNames`: convert names in the circuit to lower case
- `firrtl.features.UpperCaseNames`: convert names in the circuit to upper case

This is narrowly why I needed this transform (I need modules to be lower case).

### CLI for Lower Case / Upper Case added

This adds a new command line option for lower casing or upper casing all names:

```
  --change-name-case <lower|upper>
                           Convert all FIRRTL names to a specific case
```
### `VerilogRename` Performance

The refactoring yielded some pretty good performance improvements. Running `RenameVerilog` on `HwachaSequencer` yields the following.

Before:
```
Mean:   637.5 ms
Median: 621.4 ms
Stddev: 39.7 ms
```

After:
```
Mean:   87.7 ms
Median: 76.3 ms
Stddev: 26.8 ms
```

On a Rocket Chip `DefaultConfig`-like design:

Before:
```
Mean:   800.0 ms
Median: 779.2 ms
Stddev: 51.3 ms
```

After:
```
Mean:   192.5 ms
Median: 189.0 ms
Stddev: 12.8 ms
```

#### Other Information

This uses a transform pattern that I've been thinking about for a while and could likely be made even more abstract.

This walks the AST while tracking and updating the current `Target` scope. In effect, for whatever statement, expression, name, etc. the transform always knows what the target is. This enables the transform to trivially determine renames. Additionally, if the circuit is SSA and is traversed from module leaves to the circuit root, it allows a mutable rename map to be a store of information about whether or not anything should be renamed.

This is currently tied to low FIRRTL IR, but I think this would be a powerful pattern that could be reused a lot of places.

### Contributor Checklist

- [x] Did you add Scaladoc to every public function/method?
- [x] Did you update the FIRRTL spec to include every new feature/behavior?
- [x] Did you add at least one test demonstrating the PR?
- [x] Did you delete any extraneous printlns/debugging code?
- [x] Did you specify the type of improvement?
- [x] Did you state the API impact?
- [x] Did you specify the code generation impact?
- [x] Did you request a desired merge strategy?
- [x] Did you add text to be included in the Release Notes for this change?

#### Type of Improvement

<!-- Choose one or more from the following: -->
- bug fix
- performance improvement
<!--   - documentation                      -->
- code refactoring
<!--   - code cleanup                       -->
<!--   - backend code generation            -->
- new feature/API

#### API Impact

<!-- How would this affect the current API? Does this add, extend, deprecate, remove, or break any existing API? -->

Pure API expansion.

#### Backend Code Generation Impact

<!-- Does this change any generated Verilog?  -->
<!-- How does it change it or in what circumstances would it?  -->

None (unless a user intentionally uses one of the letter case transforms).

#### Desired Merge Strategy

<!-- If approved, how should this PR be merged? -->
<!-- Options are: -->
<!--   - Squash: The PR will be squashed and merged (choose this if you have no preference. -->
   - Rebase: You will rebase the PR onto master and it will be merged with a merge commit.

#### Release Notes
<!--
Text from here to the end of the body will be considered for inclusion in the release notes for the version containing this pull request.
-->

- Improve performance of `VerilogRename`
- Add lower and upper case name transforms

### Reviewer Checklist (only modified by reviewer)
- [ ] Did you add the appropriate labels?
- [ ] Did you mark the proper milestone (1.2.x, 1.3.0, 1.4.0) ?
- [ ] Did you review?
- [ ] Did you check whether all relevant Contributor checkboxes have been checked?
- [N/A] Did you mark as `Please Merge`?