@seldridge 2020-06-03
This is gold-plated in the extreme and ready for review.