@seldridge 2020-06-08
Discussed during the dev meeting:

1) Ask @albertchen-sifive to take a look as he's been looking at Rocket Chip renaming
2) Change the white/blacklist to something else (allow/block, skip/dontSkip, `$bikeshedColorlist`)