@seldridge 2020-06-15
Updated with feedback from dev meeting. Friendly ping @albertchen-sifive for some thoughts on how this fits within the broader naming ecosystem of Chisel and Rocket Chip.