@albertchen-sifive 2020-06-15
> Updated with feedback from dev meeting. Friendly ping @albertchen-sifive for some thoughts on how this fits within the broader naming ecosystem of Chisel and Rocket Chip.

this looks great! I think this can probably replace the renaming stuff in Rocket-Chip. I think it would be helpful if it emitted an annotation that contains the old and new module names for renamed modules (just the ones targeted by `ManipulateNamesAnnotation`, not the renames for fixing conflicts). For example, `LintConflictingModuleNames` relies on a set of names that are tracked separately from those in the `RenameMap`, so it needs a way to get this info.