@seldridge 2020-06-16
Thanks for the feedback @albertchen-sifive. I actually wasn't expecting this to be quite that universal, but if it is that's awesome.

For the name tracking, I thought about something more complex, but I think the following may just generally work. We add a `TrackerAnnotation` (could be single or multi-target). This is initialized with targets and a copy of the targets. The former are renamed while the latter stay the same. Before name manipulation starts, a tracker is dropped based on the allow list targets:

```scala
case class TrackerAnnotation(
  targets: Seq[Seq[Target]],
  oldTargets: Seq[Seq[Target]]) extends MultiTargetAnnotation {

  override def duplicate(a: Seq[Seq[Target]]) = this.copy(targets = a)

}

object TrackerAnnotation {

  def track(targets: Seq[Target]) = {
    val a = targets.map(Seq(_))
    TrackerAnnotation(a, a)
  }

}
```

This may need a custom update method or at least something to handle the edge case of `targets` shrinking. Also, there's a question of how to key into a specific tracker annotation. Additionally, a helper method to generate a rename map from `targets` and `oldTargets`.

I'll get back to this tomorrow and see what `DesiredNameAnnotation` is doing.