@seldridge 2020-06-17
@albertchen-sifive: I updated this with a `ManipulateNamesAllowlistResultAnnotation` that will store the result of anything renamed via a `ManipulateNamesAllowlistAnnotation`. Put differently, explicit requests to rename targets will cause a record of the rename to be permanently recorded. 

I *think* that this will serve your needs, but let me know.