@albertchen-sifive 2020-06-19
> I _think_ that this will serve your needs, but let me know.

yep that works. thanks!