@seldridge 2020-06-25
Thanks for reviewing @albertchen-sifive! I'll apply your suggestions with some fixup commits.

I agree that the abuse of implicit to hide parameters is an anti-pattern (and the PR has to use explicit parameters in the majority of places already). I changed that, too.

The [Databricks Style Guide](https://github.com/databricks/scala-style-guide/blob/master/README.md#recursion) suggests avoiding multiple parameter lists. I know those are used in a lot of other transforms, but I'll just switch to one parameter list with partial application.