@seldridge 2020-06-25
@albertchen-sifive: Quick re-review request for the two commits I added (primarily if you're fine with how I'm removing the implicit parameters). If you approve, I'll rebase to kill the `!fixup` commit and manually merge.