@seldridge 2020-06-25
One more commit to fix some issues with Scala 2.11 not being able to infer the type of partially applied function placeholders.